Test with large arrays [issue #10 #11]

+++ jmatrw/src/main/java/it/prz/jmatrw/io/DataElement.java
@@ -30,6 +30,6 @@ public class DataElement {
 	
 	public DEType dataElementType = DEType.STANDARD;
 	public MLDataType dataType;
-	public int numOfBytesBody;
+	public long numOfBytesBody;
 
 }//EndClass.
